Refactor file cache into own sub-module.
